+++ common/src/main/java/com/coolerpromc/resourcestrees/Constants.java
@@ -1,10 +1,20 @@
 package com.coolerpromc.resourcestrees;
 
+import com.coolerpromc.resourcestrees.api.tree.TreeTypes;
+import com.coolerpromc.resourcestrees.client.pack.ResourcesTreesModelPack;
 import net.minecraft.network.chat.MutableComponent;
 import net.minecraft.network.chat.contents.TranslatableContents;
 import net.minecraft.resources.Identifier;
+import net.minecraft.server.packs.PackLocationInfo;
+import net.minecraft.server.packs.PackResources;
+import net.minecraft.server.packs.PackSelectionConfig;
+import net.minecraft.server.packs.repository.Pack;
+import net.minecraft.server.packs.repository.PackCompatibility;
+import net.minecraft.world.flag.FeatureFlagSet;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+
+import java.util.List;
 
 public class Constants {
 	public static final String MODID = "resourcestrees";
@@ -29,5 +39,27 @@ public class Constants {
 
 	public static String getOrFallback(String translationKey, String fallback){
 		return MutableComponent.create(new TranslatableContents(translationKey, capitalizeWords(fallback), TranslatableContents.NO_ARGS)).getString();
+	}
+
+	public static Pack getInMemoryPack(){
+		ResourcesTreesModelPack pack = new ResourcesTreesModelPack();
+		TreeTypes.getTypes().forEach(treeType -> {
+			pack.addLeavesModel(treeType.leavesTexture());
+			pack.addSaplingBlockModel(treeType.saplingTexture());
+			pack.addSaplingItemModel(treeType.saplingTexture());
+		});
+
+		Pack p = new Pack(
+				pack.location(),
+				new Pack.ResourcesSupplier() {
+					@Override
+					public PackResources openPrimary(PackLocationInfo info) { return pack; }
+					@Override
+					public PackResources openFull(PackLocationInfo info, Pack.Metadata meta) { return pack; }
+				},
+				new Pack.Metadata(pack.getDescription(), PackCompatibility.COMPATIBLE, FeatureFlagSet.of(), List.of()),
+				new PackSelectionConfig(true, Pack.Position.TOP, false)
+		);
+		return p;
 	}
 }